requirements.txt
# Core dependencies
tinkoff-investments==0.3.1
psycopg2-binary==2.9.9
pandas==2.3.1
numpy==2.3.1
matplotlib==3.10.3
python-dotenv==1.1.1
openpyxl==3.1.2
tqdm==4.66.4
